requirements.txt
emoji==2.2.0
google-api-python-client==2.76.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-cloud-storage==2.7.0
nbformat==5.7.3
numpy==1.24.1
pandas==1.5.3
plotly==5.13.0
tqdm==4.64.1